examples/ernie-completion-examples: make this example a separate module
Right now it's in the main module, the only example of this kind.
Align it with the other examples.

This is a two step change, and this is step 1; step 2 will be updating
this example's go.mod to rely on a newer version of langchaingo that
no longer provides this package.