textsplitter: support reference links in markdown_splitter
The markdown splitter function 'splitInline' has been significantly
enhanced to handle a variety of inline elements as per the CommonMark
specification. Previously, it simply returned the content of the inline
element. Now, it checks for the presence of children in the inline
element and processes them accordingly.

This change was necessary to support reference links, which have the
following format:

```markdown
[foo][bar]

[bar]: /url "title"
```

When this gets parsed, the URL and title from the definition are
added to each reference link and the definition is dropped from the
document. This means when the document is rendered we end up with
incomplete links.

With this change, we render full links instead of references, so the
example above ends up looking like:

```markdown
[foo](/url "title")
```