vectorstores: Weaviate add additional fields option
Via `queryAttrs` you can specify fields to be returned with a
similarity search. In Weaviate you can retrieve the id via the
`_additional` field. With the current setup it is not possible to
add additional fields (like the id) to the the `_additional` due to
the manual mapping in 'createFields' function.

This adds a new `WithAdditionalFields` optional to give the user
control over which additional fields to return.