llms: add caching functionality for Models
New caching functionality has been added to langchaingo to boost
execution speed for repetitive tasks.  Specifically, we implemented
an in-memory caching for Language Learning Models (LLMs) so that once
a model generates a content from a sequence of messages, this result
is stored in the cache. Any subsequent calls with the same sequence of
messages will read the result from the cache, instead of having the
LLM regenerate it, this optimization is intended to reduce the
response time for recurring requests.

This caching functionality is generic, different cache backends can be
used when creating the wrapper. To accomplish this, the 'llms/cache'
package was created with a generic wrapper that adds caching to a
'llms.Model'. The 'llms/cache/inmemory' package was also created for
the in-memory implementation of the cache.

Additionally, a caching example was included to demonstrate the usage
of the implemented caching mechanism.

Minor fix: Typo error in 'TotalTokesn' was corrected to 'TotalToken'
in 'ollama/ollamallm.go'.

Resolves #395