Changed the logic with the custom ID slightly
The rationale is that you should have the option to set a custom ID for each embedding document. Therefor, I’ve added a property CustomID to the Document interface. It’s defined as a nullable string so that you can leave it as a null value if you want.

This change also required an update of the text splitter to keep the custom ID during the splitting.

I also update the pinecone integration so that it fixes https://github.com/tmc/langchaingo/issues/306 as well.