feat: add support for deprecated max_tokens open ai param
This is a cleaner method of using the deprecated `max_tokens` parameter
when using open ai compatible APIs who have not yet deprecated the
parameter.

It gives the consumer the ability to determine, prior to generating chat
content, the ability to determine whether or not they're using a
different base URL, which base URL they're using, and then make an
informed decision on whether or not to use the deprecated `max_tokens`
parameter instead of the default `max_completion_tokens` parameter.